Producto/Back-End/Vambot/discovery_service/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
google-generativeai==0.4.1
pgvector==0.2.5
python-dotenv==1.0.1
psycopg2-binary==2.9.11